fix: email send error
