+++ backend/src/main/java/com/shinhan/shbhack/ijoa/api/controller/member/EmailController.java
@@ -5,6 +5,8 @@ import com.shinhan.shbhack.ijoa.api.service.member.dto.request.EmailCheckService
 import com.shinhan.shbhack.ijoa.api.service.member.query.EmailQueryService;
 import com.shinhan.shbhack.ijoa.api.service.member.query.MemberQueryService;
 import com.shinhan.shbhack.ijoa.common.response.ApiData;
+import com.shinhan.shbhack.ijoa.common.util.EmailUtil;
+import com.shinhan.shbhack.ijoa.common.util.RedisUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
@@ -20,13 +22,18 @@ import java.io.UnsupportedEncodingException;
 @RequiredArgsConstructor
 public class EmailController {
 
-    private final MemberQueryService memberQueryService;
     private final EmailQueryService emailQueryService;
+    private final EmailUtil emailUtil;
+    private final RedisUtil redisUtil;
 
     @PostMapping("/send")
     public ApiData<String> sendJoinEmail(@RequestParam("email") @Email String email) throws MessagingException, UnsupportedEncodingException {
         // TODO: 2023-09-11 need refactoring
-        emailQueryService.sendEmail(email);
+
+        String code = emailQueryService.sendEmail(email);
+
+        emailUtil.joinEmail(email, code);
+        redisUtil.setEmail(email, code);
 
         return ApiData.of("이메일이 발송되었습니다.");
     }
