+++ backend/src/main/java/com/shinhan/shbhack/ijoa/api/service/member/query/EmailQueryService.java
@@ -29,24 +29,15 @@ import java.util.Random;
 @RequiredArgsConstructor
 public class EmailQueryService {
 
-    private final JavaMailSender emailSender;
     private final MemberRepository memberRepository;
     private final RedisUtil redisUtil;
 
-    public void sendEmail(String email) throws MessagingException, UnsupportedEncodingException {
+    public String sendEmail(String email) throws MessagingException, UnsupportedEncodingException {
         if(memberRepository.existsByEmail(email))
             throw new EntityNotFoundException(ErrorCode.MEMBER_DUPLICATE);
 
-        String code = createCode();
+        return createCode();
 
-        try {
-            MimeMessage emailForm = createEmailForm(email, code);
-            emailSender.send(emailForm);
-        }catch (Exception e){
-            throw new InvalidValueException(ErrorCode.EMAIL_FORM_ERROR);
-        }
-
-        redisUtil.setEmail(email, code);
     }
 
     private String createCode() {
@@ -64,32 +55,6 @@ public class EmailQueryService {
         }
     }
 
-    // 발신할 이메일 데이터 세팅
-    private MimeMessage createEmailForm(String email, String code) throws MessagingException, UnsupportedEncodingException {
-        String senderEmail = "noreply@moailgi.com"; // Replace with your email address (sender)
-        String senderName = "모아일기"; // Replace with your name (sender)
-
-        MimeMessage message = emailSender.createMimeMessage();
-        MimeMessageHelper helper = new MimeMessageHelper(message, false, "utf-8");
-        helper.setTo(email); // Set recipient email address
-        helper.setFrom(new InternetAddress(senderEmail, senderName)); // Set sender email and name
-        helper.setSubject("Verification Code for Your Account"); // Set email subject
-
-        // Email content with the generated verification code
-        String emailContent = "<html><body style=\"font-family: Arial, sans-serif;\">"
-                + "<h2>안녕하세요!!</h2>"
-                + "<p>모아일기 사이트에 회원가입을 해주셔서 감사합니다!</p>"
-                + "<p>인증코드입니다.:</p>"
-                + "<h3 style=\"background-color: #f0f0f0; padding: 10px;\">" + code + "</h3>"
-                + "<p>Please use this code to verify your account.</p>"
-                + "<p>Best regards,<br/>Your Website Team</p>"
-                + "</body></html>";
-
-        helper.setText(emailContent, true); // Set email content as HTML
-
-        return message;
-    }
-
     public void checkEmail(EmailCheckServiceRequest request){
         Optional<String> code = redisUtil.getEmail(request.getEmail());
         if(!code.isPresent()) throw new InvalidValueException(ErrorCode.NOTMATCH_EMAIL);
@@ -97,5 +62,6 @@ public class EmailQueryService {
 
         redisUtil.deleteEmail(request.getEmail());
     }
+
 
 }
