+++ backend/src/main/java/com/shinhan/shbhack/ijoa/common/util/RedisUtil.java
@@ -5,13 +5,14 @@ import com.shinhan.shbhack.ijoa.common.model.UserDetailsModel;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 
 import java.time.Duration;
 import java.util.Optional;
 
 @Slf4j
-@Repository
+@Component
 @RequiredArgsConstructor
 public class RedisUtil {
